- Cleaning up and removing redundant code duplicated from tasks.py
